requirements.txt
asgiref==3.2.3
boto3==1.12.39
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
django-email-verification==0.0.7
django-filter==2.2.0
djangorestframework==3.11.0
django-rest-passwordreset==1.1.0
idna==2.8
lxml==4.5.0
Markdown==3.1.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.15.0
sqlparse==0.3.0
unicodecsv==0.14.1
urllib3==1.25.8
